♻️ refactor(manage): split Settings into separate Settings and Management pages
Separated the single Settings page into two distinct pages:
- /settings → Ajustes (General + Admin: Usuarios)
- /manage → Gestionar Contenido (Materias, Unidades, Preguntas)

The Gestionar button in the navbar now navigates to /manage,
while the Ajustes link in the user dropdown keeps pointing to /settings.
Extracted Management component from Settings to keep concerns separated.
