+++ frontend/src/pages/SettingsPage.tsx
@@ -1,5 +1,5 @@
 import Settings from '../components/Settings';
 
-export default function SettingsPage({ isAdmin, token, onSubjectsChanged }: { isAdmin: boolean; token: string; onSubjectsChanged?: () => void }) {
-  return <Settings isAdmin={isAdmin} token={token} onSubjectsChanged={onSubjectsChanged} />;
+export default function SettingsPage({ isAdmin, token }: { isAdmin: boolean; token: string }) {
+  return <Settings isAdmin={isAdmin} token={token} />;
 }
